refactor: modal dialog open으로 변경
