+++ frontend/src/hooks/useModal.ts
@@ -1,7 +1,8 @@
-import { useState } from 'react';
+import { useEffect, useRef, useState } from 'react';
 
 const useModal = (initialState = false) => {
   const [isOpen, setIsOpen] = useState(initialState);
+  const modalRef = useRef<HTMLDialogElement>(null);
 
   const on = () => {
     setIsOpen(true);
@@ -16,7 +17,30 @@ const useModal = (initialState = false) => {
     setTimeout(off, ms);
   };
 
-  return { isOpen, on, off, onTime };
+  const keyDownHandler = (e: KeyboardEvent) => {
+    if (e.key === 'Escape') {
+      off();
+    }
+  };
+  const body = useRef(document.body);
+
+  useEffect(() => {
+    if (isOpen) {
+      modalRef.current?.showModal();
+      body.current.querySelector('#root')?.setAttribute('aria-hidden', 'true');
+      body.current.style.overflowY = 'hidden';
+      window.addEventListener('keydown', keyDownHandler);
+    }
+
+    return () => {
+      modalRef.current?.close();
+      body.current.querySelector('#root')?.setAttribute('aria-hidden', 'false');
+      body.current.style.overflowY = 'auto';
+      window.removeEventListener('keydown', keyDownHandler);
+    };
+  }, [isOpen]);
+
+  return { isOpen, on, off, onTime, modalRef };
 };
 
 export default useModal;
